+++ ckb_syscalls.h
@@ -8,6 +8,10 @@
 #include "ckb_consts.h"
 #include "ckb_syscall_apis.h"
 
+#ifndef CKB_STDLIB_NO_SYSCALL_IMPL
+
+#include "ckb_raw_syscalls.h"
+
 int ckb_checked_load_tx_hash(void* addr, uint64_t* len, size_t offset) {
   uint64_t old_len = *len;
   int ret = ckb_load_tx_hash(addr, len, offset);
@@ -125,10 +129,6 @@ int ckb_checked_load_cell_data(void* addr, uint64_t* len, size_t offset,
   return ret;
 }
 
-/* load the actual witness for the current type verify group.
-   use this instead of ckb_load_witness if type contract needs args to verify
-   input/output.
- */
 int ckb_load_actual_type_witness(uint8_t* buf, uint64_t* len, size_t index,
                                  size_t* type_source) {
   *type_source = CKB_SOURCE_GROUP_INPUT;
@@ -142,7 +142,6 @@ int ckb_load_actual_type_witness(uint8_t* buf, uint64_t* len, size_t index,
   return ckb_checked_load_witness(buf, len, 0, index, *type_source);
 }
 
-/* calculate inputs length */
 int ckb_calculate_inputs_len() {
   uint64_t len = 0;
   /* lower bound, at least tx has one input */
@@ -179,10 +178,6 @@ int ckb_calculate_inputs_len() {
   return hi;
 }
 
-/*
- * Look for a dep cell with specific code hash, code_hash should be a buffer
- * with 32 bytes.
- */
 int ckb_look_for_dep_with_hash2(const uint8_t* code_hash, uint8_t hash_type,
                                 size_t* index) {
   size_t current = 0;
@@ -216,155 +211,6 @@ int ckb_look_for_dep_with_hash(const uint8_t* data_hash, size_t* index) {
   return ckb_look_for_dep_with_hash2(data_hash, 0, index);
 }
 
-#ifndef CKB_STDLIB_NO_SYSCALL_IMPL
-
-#define memory_barrier() asm volatile("fence" ::: "memory")
-
-static inline long __internal_syscall(long n, long _a0, long _a1, long _a2,
-                                      long _a3, long _a4, long _a5) {
-  register long a0 asm("a0") = _a0;
-  register long a1 asm("a1") = _a1;
-  register long a2 asm("a2") = _a2;
-  register long a3 asm("a3") = _a3;
-  register long a4 asm("a4") = _a4;
-  register long a5 asm("a5") = _a5;
-
-#ifdef __riscv_32e
-  register long syscall_id asm("t0") = n;
-#else
-  register long syscall_id asm("a7") = n;
-#endif
-
-  asm volatile("scall"
-               : "+r"(a0)
-               : "r"(a1), "r"(a2), "r"(a3), "r"(a4), "r"(a5), "r"(syscall_id));
-  /*
-   * Syscalls might modify memory sent as pointer, adding a barrier here ensures
-   * gcc won't do incorrect optimization.
-   */
-  memory_barrier();
-
-  return a0;
-}
-
-#define syscall(n, a, b, c, d, e, f)                                           \
-  __internal_syscall(n, (long)(a), (long)(b), (long)(c), (long)(d), (long)(e), \
-                     (long)(f))
-
-int ckb_exit(int8_t code) { return syscall(SYS_exit, code, 0, 0, 0, 0, 0); }
-
-int ckb_load_tx_hash(void* addr, uint64_t* len, size_t offset) {
-  volatile uint64_t inner_len = *len;
-  int ret = syscall(SYS_ckb_load_tx_hash, addr, &inner_len, offset, 0, 0, 0);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_load_script_hash(void* addr, uint64_t* len, size_t offset) {
-  volatile uint64_t inner_len = *len;
-  int ret =
-      syscall(SYS_ckb_load_script_hash, addr, &inner_len, offset, 0, 0, 0);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_load_cell(void* addr, uint64_t* len, size_t offset, size_t index,
-                  size_t source) {
-  volatile uint64_t inner_len = *len;
-  int ret =
-      syscall(SYS_ckb_load_cell, addr, &inner_len, offset, index, source, 0);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_load_input(void* addr, uint64_t* len, size_t offset, size_t index,
-                   size_t source) {
-  volatile uint64_t inner_len = *len;
-  int ret =
-      syscall(SYS_ckb_load_input, addr, &inner_len, offset, index, source, 0);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_load_header(void* addr, uint64_t* len, size_t offset, size_t index,
-                    size_t source) {
-  volatile uint64_t inner_len = *len;
-  int ret =
-      syscall(SYS_ckb_load_header, addr, &inner_len, offset, index, source, 0);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_load_witness(void* addr, uint64_t* len, size_t offset, size_t index,
-                     size_t source) {
-  volatile uint64_t inner_len = *len;
-  int ret =
-      syscall(SYS_ckb_load_witness, addr, &inner_len, offset, index, source, 0);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_load_script(void* addr, uint64_t* len, size_t offset) {
-  volatile uint64_t inner_len = *len;
-  int ret = syscall(SYS_ckb_load_script, addr, &inner_len, offset, 0, 0, 0);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_load_transaction(void* addr, uint64_t* len, size_t offset) {
-  volatile uint64_t inner_len = *len;
-  int ret =
-      syscall(SYS_ckb_load_transaction, addr, &inner_len, offset, 0, 0, 0);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_load_cell_by_field(void* addr, uint64_t* len, size_t offset,
-                           size_t index, size_t source, size_t field) {
-  volatile uint64_t inner_len = *len;
-  int ret = syscall(SYS_ckb_load_cell_by_field, addr, &inner_len, offset, index,
-                    source, field);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_load_header_by_field(void* addr, uint64_t* len, size_t offset,
-                             size_t index, size_t source, size_t field) {
-  volatile uint64_t inner_len = *len;
-  int ret = syscall(SYS_ckb_load_header_by_field, addr, &inner_len, offset,
-                    index, source, field);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_load_input_by_field(void* addr, uint64_t* len, size_t offset,
-                            size_t index, size_t source, size_t field) {
-  volatile uint64_t inner_len = *len;
-  int ret = syscall(SYS_ckb_load_input_by_field, addr, &inner_len, offset,
-                    index, source, field);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_load_cell_data(void* addr, uint64_t* len, size_t offset, size_t index,
-                       size_t source) {
-  volatile uint64_t inner_len = *len;
-  int ret = syscall(SYS_ckb_load_cell_data, addr, &inner_len, offset, index,
-                    source, 0);
-  *len = inner_len;
-  return ret;
-}
-
-int ckb_debug(const char* s) {
-  return syscall(SYS_ckb_debug, s, 0, 0, 0, 0, 0);
-}
-
-int ckb_vm_version() { return syscall(SYS_ckb_vm_version, 0, 0, 0, 0, 0, 0); }
-
-uint64_t ckb_current_cycles() {
-  return syscall(SYS_ckb_current_cycles, 0, 0, 0, 0, 0, 0);
-}
-
 int ckb_exec_cell(const uint8_t* code_hash, uint8_t hash_type, uint32_t offset,
                   uint32_t length, int argc, const char* argv[]) {
   size_t index = SIZE_MAX;
@@ -373,13 +219,7 @@ int ckb_exec_cell(const uint8_t* code_hash, uint8_t hash_type, uint32_t offset,
     return ret;
   }
   size_t bounds = ((size_t)offset << 32) | length;
-  return syscall(SYS_ckb_exec, index, CKB_SOURCE_CELL_DEP, 0, bounds, argc,
-                 argv);
-}
-
-int ckb_spawn(size_t index, size_t source, size_t place, size_t bounds,
-              spawn_args_t* spawn_args) {
-  return syscall(SYS_ckb_spawn, index, source, place, bounds, spawn_args, 0);
+  return ckb_exec(index, CKB_SOURCE_CELL_DEP, 0, bounds, argc, argv);
 }
 
 int ckb_spawn_cell(const uint8_t* code_hash, uint8_t hash_type, uint32_t offset,
@@ -390,50 +230,8 @@ int ckb_spawn_cell(const uint8_t* code_hash, uint8_t hash_type, uint32_t offset,
     return ret;
   }
   size_t bounds = ((size_t)offset << 32) | length;
-  return syscall(SYS_ckb_spawn, index, CKB_SOURCE_CELL_DEP, 0, bounds,
-                 spawn_args, 0);
+  return ckb_spawn(index, CKB_SOURCE_CELL_DEP, 0, bounds, spawn_args);
 }
-
-int ckb_wait(uint64_t pid, int8_t* exit_code) {
-  return syscall(SYS_ckb_wait, pid, exit_code, 0, 0, 0, 0);
-}
-
-uint64_t ckb_process_id() {
-  return syscall(SYS_ckb_process_id, 0, 0, 0, 0, 0, 0);
-}
-
-int ckb_pipe(uint64_t fds[2]) {
-  return syscall(SYS_ckb_pipe, fds, 0, 0, 0, 0, 0);
-}
-
-int ckb_read(uint64_t fd, void* buffer, size_t* length) {
-  volatile size_t l = *length;
-  int ret = syscall(SYS_ckb_read, fd, buffer, &l, 0, 0, 0);
-  *length = l;
-  return ret;
-}
-
-int ckb_write(uint64_t fd, const void* buffer, size_t* length) {
-  volatile size_t l = *length;
-  int ret = syscall(SYS_ckb_write, fd, buffer, &l, 0, 0, 0);
-  *length = l;
-  return ret;
-}
-
-int ckb_load_block_extension(void* addr, uint64_t* len, size_t offset,
-                             size_t index, size_t source) {
-  return syscall(SYS_ckb_load_block_extension, addr, len, offset, index, source,
-                 0);
-}
-
-int ckb_inherited_fds(uint64_t* fds, size_t* length) {
-  volatile size_t l = *length;
-  int ret = syscall(SYS_ckb_inherited_fds, fds, &l, 0, 0, 0, 0);
-  *length = l;
-  return ret;
-}
-
-int ckb_close(uint64_t fd) { return syscall(SYS_ckb_close, fd, 0, 0, 0, 0, 0); }
 
 #endif /* CKB_STDLIB_NO_SYSCALL_IMPL */
